requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
attrs
blinker
certifi
click
Flask
frozenlist
idna
itsdangerous
Jinja2
MarkupSafe
multidict
propcache
pyee
python-dotenv
slackclient
slackeventsapi
typing_extensions
Werkzeug
yarl
